+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ### Patch Changes
 
+- Updated dependencies [3bd2689]
+  - ai@6.0.0-beta.139
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - ai@6.0.0-beta.138
 
 ## 0.0.1-beta.0